BUGFIX: Logic for allocating bonus Maneuvers
Fixed a bug in the logic for allocating bonus Maneuvers to Complexity-
based bonus slots, where it disregarded the Complexity of the granted
slot.

Also added a tracking property to visually display on the character
sheet when a Maneuver is marked to be handled by bonus Complexity
but there wasn't a valid, available slot for it.